Added new response entities to Artist Controller

+++ src/main/java/pl/jakub/musiclibrary/artist/ArtistController.java
@@ -6,6 +6,7 @@ import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
+import java.util.Optional;
 
 @RestController
 @RequestMapping("/artists")
@@ -19,22 +20,44 @@ public class ArtistController {
 
     @GetMapping
     public ResponseEntity<List<Artist>> getAll() {
-        return new ResponseEntity<>(artistService.findAll(), HttpStatus.OK);
+        List<Artist> allArtists = artistService.findAll();
+        if (allArtists == null)
+            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+        else if (allArtists.isEmpty())
+            return new ResponseEntity<>(allArtists, HttpStatus.NO_CONTENT);
+        else
+            return new ResponseEntity<>(allArtists, HttpStatus.OK);
     }
 
     @GetMapping("/page")
     public ResponseEntity<List<Artist>> getSortedPage(@RequestParam Integer pageNumber, @RequestParam Integer pageSize, @RequestParam String sort, @RequestParam(defaultValue = "false") Boolean descending) {
-        return new ResponseEntity<>(artistService.findSortedPage(pageNumber, pageSize, sort, descending), HttpStatus.OK);
+        List<Artist> page = artistService.findSortedPage(pageNumber, pageSize, sort, descending);
+        if (page == null)
+            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+        else if (page.isEmpty())
+            return new ResponseEntity<>(page, HttpStatus.NO_CONTENT);
+        else
+            return new ResponseEntity<>(page, HttpStatus.OK);
     }
 
     @GetMapping("/{id}")
     public ResponseEntity<Artist> getById(@PathVariable Long id) {
-        return new ResponseEntity<>(artistService.findById(id), HttpStatus.OK);
+        Artist artist = artistService.findById(id);
+        if (artist == null)
+            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+        else
+            return new ResponseEntity<>(artist, HttpStatus.OK);
     }
 
     @GetMapping("/by-name")
     public ResponseEntity<List<Artist>> getByName(@RequestParam String name) {
-        return new ResponseEntity<>(artistService.findByName(name), HttpStatus.OK);
+        List<Artist> artists = artistService.findByName(name);
+        if (artists == null)
+            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+        else if (artists.isEmpty())
+            return new ResponseEntity<>(artists, HttpStatus.NO_CONTENT);
+        else
+            return new ResponseEntity<>(artists, HttpStatus.OK);
     }
 
     @PostMapping
@@ -44,12 +67,21 @@ public class ArtistController {
 
     @PutMapping("/{id}")
     public ResponseEntity<Artist> update(@PathVariable Long id, @Valid @RequestBody Artist updatedArtist) {
-        return new ResponseEntity<>(artistService.update(id, updatedArtist), HttpStatus.OK);
+        Artist artist = artistService.findById(id);
+        if (artist == null)
+            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+        else
+            return new ResponseEntity<>(artistService.update(id, updatedArtist), HttpStatus.OK);
     }
 
     @DeleteMapping("/{id}")
     public ResponseEntity<HttpStatus> delete(@PathVariable Long id) {
-        artistService.delete(id);
-        return new ResponseEntity<>(HttpStatus.NO_CONTENT);
+        Artist artist = artistService.findById(id);
+        if (artist == null)
+            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+        else {
+            artistService.delete(id);
+            return new ResponseEntity<>(HttpStatus.NO_CONTENT);
+        }
     }
 }
